+++ src/main/java/com/gamemetricbackend/global/config/WebSecurityConfig.java
@@ -70,8 +70,8 @@ public class WebSecurityConfig implements WebMvcConfigurer {
                 .requestMatchers(HttpMethod.GET,"/users/search").permitAll()
                 .requestMatchers(HttpMethod.GET,"/broadcasts").permitAll()
                 .requestMatchers(HttpMethod.GET,"/broadcasts/**").permitAll()
-                .requestMatchers("/signaling/**").permitAll()
-                .requestMatchers("/signaling").permitAll() // 임시 websocket 테스트용
+                .requestMatchers("/ws").permitAll()
+                .requestMatchers("/ws/**").permitAll() // 임시 websocket 테스트용
                 .requestMatchers(HttpMethod.GET,"/catagory").permitAll()
                 .requestMatchers("/v3/api-docs/**", "/swagger-ui/**", "/swagger-resources/**").permitAll()
                 .anyRequest().authenticated() // 그 외 모든 요청 인증처리
