+++ src/main/java/com/gamemetricbackend/global/config/WebSocketConfig.java
@@ -1,25 +1,24 @@
 package com.gamemetricbackend.global.config;
 
+import com.gamemetricbackend.global.websocket.WebSocketChatHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.*;
 
 @Configuration
-@EnableWebSocketMessageBroker
+@EnableWebSocket
 @RequiredArgsConstructor
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
+public class WebSocketConfig implements WebSocketConfigurer {
+
+    private final WebSocketChatHandler webSocketHandler;
 
     @Override
-    public void configureMessageBroker(MessageBrokerRegistry config) {
-        config.enableSimpleBroker("/topic"); // broker url 설정
-        config.setApplicationDestinationPrefixes("/app"); // send url 설정
-    }
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/signaling")// webSokcet 접속시 endpoint 설정
-                .setAllowedOriginPatterns("*") // cors 에 따른 설정 ( * 는 모두 허용 )
-                .withSockJS(); // 브라우저에서 WebSocket 을 지원하지 않는 경우에 대안으로 어플리케이션의 코드를 변경할 필요 없이 런타임에 필요할 때 대체하기 위해 설정
+    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
+        registry
+                // /ws/conn 경로로 WebSocket 연결을 허용
+                .addHandler(webSocketHandler, "/ws")
+                // CORS 허용
+                .setAllowedOrigins("*");
     }
 }
